requirements.txt
# Duplicate Image Finder Dependencies

# Core duplicate detection
imagededup>=0.3.2

# GUI Framework
PyQt6>=6.6.0

# Image processing
Pillow>=10.0.0
numpy>=1.24.0

# RAW file support
rawpy>=0.19.0

# Testing
pytest>=7.4.0
pytest-qt>=4.2.0
